REFACTOR: Logout buttom
