+++ src/components/index.jsx
@@ -1,3 +1,4 @@
 import { ReadMore } from "./ReadMore";
+import { Logout } from "./Logout";
 
-export { ReadMore };
+export { ReadMore, Logout };
