+++ src/views/dashboard/DashboardView.jsx
@@ -1,36 +1,37 @@
-import { Button, Center, Flex, Heading, useToast } from "@chakra-ui/react";
+import { Button, Flex, Heading, useToast } from "@chakra-ui/react";
 import React from "react";
 import { Outlet, useNavigate } from "react-router";
+import { Logout } from "../../components";
 import { useAuth } from "../../services/auth";
 
 export function DashboardView() {
-  const auth = useAuth();
-  const toast = useToast();
-  const navigate = useNavigate();
-  async function handleLogout(e) {
-    auth.setIsLoading(true);
-    try {
-      await auth.logout();
-      toast({
-        title: "Success",
-        description: "You have successfully logged out",
-        status: "success",
-        duration: 9000,
-        isClosable: true,
-      });
-    } catch (error) {
-      toast({
-        title: "Logout failed",
-        description: "something went wrong",
-        status: "error",
-        duration: 9000,
-        isClosable: true,
-      });
-      auth.setUser(null);
-      navigate("/");
-    }
-    auth.setIsLoading(false);
-  }
+  // const auth = useAuth();
+  // const toast = useToast();
+  // const navigate = useNavigate();
+  // async function handleLogout(e) {
+  //   auth.setIsLoading(true);
+  //   try {
+  //     await auth.logout();
+  //     toast({
+  //       title: "Success",
+  //       description: "You have successfully logged out",
+  //       status: "success",
+  //       duration: 9000,
+  //       isClosable: true,
+  //     });
+  //   } catch (error) {
+  //     toast({
+  //       title: "Logout failed",
+  //       description: "something went wrong",
+  //       status: "error",
+  //       duration: 9000,
+  //       isClosable: true,
+  //     });
+  //     auth.setUser(null);
+  //     navigate("/");
+  //   }
+  //   auth.setIsLoading(false);
+  // }
   // const handleLogout = (e) => {
   //   auth.logout();
   //   // navigate("/dashboard");
@@ -44,14 +45,7 @@ export function DashboardView() {
         background={"#f0f0f0"}
       >
         <Heading>Dashboard</Heading>
-        <Button
-          colorScheme="red"
-          onClick={handleLogout}
-          type="submit"
-          isLoading={auth.isLoading}
-        >
-          Logout
-        </Button>
+        <Logout />
       </Flex>
 
       <Outlet />
