--- src/views/dashboard/Logout.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export function Logout() {
-  return (
-    <>
-      <h1>Logout</h1>
-    </>
-  );
-}
