+++ src/views/dashboard/index.jsx
@@ -1,5 +1,4 @@
 import { DashboardView } from "./DashboardView";
 import { Profile, EditProfile } from "./Profile";
-import { Logout } from "./Logout";
 
-export { DashboardView, Profile, EditProfile, Logout };
+export { DashboardView, Profile, EditProfile };
